requirements.txt
BeautifulSoup==3.2.1
Django==1.5.4
Markdown==2.3.1
MySQL-python==1.2.4
PIL==1.1.7
South==0.8.2
boto==2.13.3
coverage==3.6
coveralls==0.2
django-bootstrap-toolkit==2.15.0
django-flatblocks==0.7.1
django-grappelli==2.4.6
django-haystack==2.1.0
django-nose==1.2
django-storages==1.1.8
djangorestframework==2.3.8 
feedparser==5.1.3
nose==1.3.0
pep8==1.4.6
pyelasticsearch==0.6
python-memcached==1.53
sorl-thumbnail==11.12
